+++ src/interfaces/change-log.ts
@@ -1,8 +1,6 @@
 export default interface IChangelog {
   changelogFileName: string
   newLog: string
-  newComments?: string
   logFind: string
-  commentFind: string
   encoding: any
 }