+++ src/interfaces/update-change-log.ts
@@ -1,0 +1,11 @@
+import {Context} from '@actions/github/lib/context'
+
+export default interface IUpdateChangelog {
+  toolkit: any
+  context: Context
+  changelogFileName: string
+  newLog: string
+  logFind: string
+  oldLogs: string
+  quantityLogs: number
+}