+++ src/useCases/create-new-release.ts
@@ -1,20 +1,12 @@
 import {ICreateNewRelease} from '../interfaces'
 
-function githubToken(): string {
-  const token = process.env.GITHUB_TOKEN
-  if (!token)
-    throw ReferenceError('No token defined in the environment variables')
-  return token
-}
-
 export default async function createNewRelease({
-  getOctokit,
+  toolkit,
   context,
   sha,
   logsSplit
 }: ICreateNewRelease): Promise<void> {
   try {
-    const toolkit = getOctokit(githubToken())
     const ref = `refs/heads/release/v${logsSplit[0].split('v')[1]}`
 
     await toolkit.rest.git.createRef({