+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 # v1.1.0
 
 ## Alterações
+- [DEVOPS] Add update changelog in action file
 - [DEVOPS] Update changelog action
 - [DEVOPS] Update list
 - [DEVOPS] Change comment find to log find